backend/requirements.txt
livekit-agents==1.2.5
livekit-plugins-silero
livekit-plugins-google==1.2.5
livekit-plugins-noise-cancellation
livekit-plugins-turn_detector
python-dotenv
livekit-api
google-generativeai
google-genai==1.24.0
langfuse
opentelemetry-sdk
opentelemetry-api
quart
quart-cors
cloud-sql-python-connector
pg8000
deprecated
psycopg2-binary
asyncpg
google-cloud-storage
sqlalchemy
httpx
